+++ lib/maze/client/appium/bs_client.rb
@@ -68,7 +68,7 @@ module Maze
         def log_run_intro
           # Log a link to the BrowserStack session search dashboard
           url = "https://app-automate.browserstack.com/dashboard/v2/search?query=#{Maze.run_uuid}&type=builds"
-          $logger.info Maze::LogUtil.linkify(url, 'BrowserStack session(s)')
+          $logger.info Maze::Loggers::LogUtil.linkify(url, 'BrowserStack session(s)')
         end
 
         def log_run_outro